+++ ReviewYourself/Models/Services/Implementations/TaskService.cs
@@ -7,7 +7,7 @@ namespace ReviewYourself.Models.Services.Implementations
     public class TaskService : ITaskService
     {
         private readonly ITaskRepository _taskRepository;
-        private ITokenRepository _tokenRepository;
+        private readonly ITokenRepository _tokenRepository;
 
         public TaskService(ITaskRepository taskRepository, ITokenRepository tokenRepository)
         {
@@ -17,21 +17,43 @@ namespace ReviewYourself.Models.Services.Implementations
 
         public void CreateTask(Token token, ResourceTask task)
         {
+            if (_tokenRepository.ValidateToken(token) == false)
+            {
+                throw new Exception("Wrong token info");
+            }
+            //TODO: request admins right
+
             _taskRepository.Create(task);
         }
 
         public ResourceTask GetTask(Token token, Guid taskId)
         {
+            if (_tokenRepository.ValidateToken(token) == false)
+            {
+                throw new Exception("Wrong token info");
+            }
+
             return _taskRepository.Read(taskId);
         }
 
         public ICollection<ResourceTask> GetTaskByCourse(Token token, Guid courseId)
         {
+            if (_tokenRepository.ValidateToken(token) == false)
+            {
+                throw new Exception("Wrong token info");
+            }
+
             return _taskRepository.ReadByCourse(courseId);
         }
 
         public void DeleteTask(Token token, Guid taskId)
         {
+            if (_tokenRepository.ValidateToken(token) == false)
+            {
+                throw new Exception("Wrong token info");
+            }
+            //TODO: request admins right
+
             _taskRepository.Delete(taskId);
         }
     }
